requirement.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
Appium-Python-Client==1.1.0
atomicwrites==1.4.0
attrs==20.2.0
cachetools==4.2.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect==1.1.0
cssutils==2.3.0
EasyProcess==0.3
entrypoint2==0.2.1
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lxml==4.6.3
MarkupSafe==1.1.1
more-itertools==8.5.0
mss==6.0.0
openpyxl==3.0.5
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
premailer==3.9.0
py==1.9.0
PyMySQL==0.10.0
pyparsing==2.4.7
pyscreenshot==2.2
pytest==6.2.4
pytest-assume==2.4.2
pytest-base-url==1.4.2
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-selenium==2.0.0
pytest-variables==1.9.0
PyYAML==5.3.1
requests==2.25.1
scripts==2.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.19
stua==0.2
tenacity==6.2.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
yagmail==0.14.256
zipp==3.1.0
